+++ NEWS.md
@@ -5,6 +5,7 @@
 * Replace `magrittr` import by `dplyr::%>%`.
 * Reduce the number of exported functions.
 * Move utility functions to 'utilities.R'.
+* Allow vignettes to run without MEME suite installed.
 
 
 # MotifPeeker 0.99.4